Basic 64-bit syscalls
- Fixed syscall handling to work correctly with 64-bit return values
  (for stuff like handles)

- Support for loading statically linked 64-bit ELF as root server
  binaries (searching in a tar ramdisk)

- Basic implementation of syscalls (using SYSCALL/SYSRET instructions);
  it seems to work for the most part, we use some assembly stubs in
  libsystem to translate calling conventions, this probably could be
  done much faster (by avoiding the translation) just by re-arranging
  the order of arguments for syscalls.

- Change GDT layout to support SYSCALL; this required adding an extra
  user code entry

- Minor bugfixes across the board for 64-bit support in libc/libsystem

- Print FS/GS register bases in exception dump

+++ kernel/arch/x86_64/src/syscall/ThreadLocal.cpp
@@ -24,7 +24,7 @@ using namespace arch;
  * XXX: What validation, if any, do we need to do on the base address? Seems like paging should
  * protect us from everything.
  */
-int syscall::UpdateThreadTlsBase(const ::sys::Syscall::Args *args, const uintptr_t) {
+intptr_t syscall::UpdateThreadTlsBase(const ::sys::Syscall::Args *args, const uintptr_t) {
     auto thread = sched::Thread::current();
 
     // get the thread
