+++ kernel/include/arch.h
@@ -108,7 +108,7 @@ void ReturnToUser(const uintptr_t pc, const uintptr_t stack, const uintptr_t arg
 /**
  * Handles architecture-specific system call.
  */
-int HandleSyscall(const sys::Syscall::Args *args, const uintptr_t number);
+intptr_t HandleSyscall(const sys::Syscall::Args *args, const uintptr_t number);
 
 /**
  * Tick callback
