+++ kernel/src/sys/Handlers.h
@@ -10,58 +10,58 @@
 namespace sys {
 
 /// Sends a message to the given port.
-int PortSend(const Syscall::Args *args, const uintptr_t number);
+intptr_t PortSend(const Syscall::Args *args, const uintptr_t number);
 /// Waits to receive a message on a port.
-int PortReceive(const Syscall::Args *args, const uintptr_t number);
+intptr_t PortReceive(const Syscall::Args *args, const uintptr_t number);
 /// Updates a port's queue size
-int PortSetParams(const Syscall::Args *args, const uintptr_t number);
+intptr_t PortSetParams(const Syscall::Args *args, const uintptr_t number);
 /// Allocates a new port.
-int PortAlloc(const Syscall::Args *args, const uintptr_t number);
+intptr_t PortAlloc(const Syscall::Args *args, const uintptr_t number);
 /// Releases a previously allocated port.
-int PortDealloc(const Syscall::Args *args, const uintptr_t number);
+intptr_t PortDealloc(const Syscall::Args *args, const uintptr_t number);
 
 /// Sets the given notification bits in the thread.
-int NotifySend(const Syscall::Args *args, const uintptr_t number);
+intptr_t NotifySend(const Syscall::Args *args, const uintptr_t number);
 /// Blocks waiting to receive a notification.
-int NotifyReceive(const Syscall::Args *args, const uintptr_t number);
+intptr_t NotifyReceive(const Syscall::Args *args, const uintptr_t number);
 
 /// Allocates a virtual memory region backed by physical memory
-int VmAlloc(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmAlloc(const Syscall::Args *args, const uintptr_t number);
 /// Allocate a virtual memory region backed by anonymous memory
-int VmAllocAnon(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmAllocAnon(const Syscall::Args *args, const uintptr_t number);
 /// Update permissions (R/W/X flags) of a VM region
-int VmRegionUpdatePermissions(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmRegionUpdatePermissions(const Syscall::Args *args, const uintptr_t number);
 /// Resizes a VM region
-int VmRegionResize(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmRegionResize(const Syscall::Args *args, const uintptr_t number);
 /// Maps a VM region into a task.
-int VmRegionMap(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmRegionMap(const Syscall::Args *args, const uintptr_t number);
 /// Unmaps a VM region.
-int VmRegionUnmap(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmRegionUnmap(const Syscall::Args *args, const uintptr_t number);
 /// Gets information about a VM region
-int VmRegionGetInfo(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmRegionGetInfo(const Syscall::Args *args, const uintptr_t number);
 /// Gets information on a task's virtual memory environment
-int VmTaskGetInfo(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmTaskGetInfo(const Syscall::Args *args, const uintptr_t number);
 /// Returns the handle for the region containing an address.
-int VmAddrToRegion(const Syscall::Args *args, const uintptr_t number);
+intptr_t VmAddrToRegion(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Creates a new userspace thread.
  */
-int ThreadCreate(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadCreate(const Syscall::Args *args, const uintptr_t number);
 /**
  * Destroys an userspace thread.
  */
-int ThreadDestroy(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadDestroy(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Sets the thread priority.
  */
-int ThreadSetPriority(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadSetPriority(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Sets the thread's notification mask.
  */
-int ThreadSetNoteMask(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadSetNoteMask(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Sets the descriptive name of the thread.
@@ -69,17 +69,17 @@ int ThreadSetNoteMask(const Syscall::Args *args, const uintptr_t number);
  * The first argument is a thread handle (or zero for current thread,) followed by the userspace
  * address of the name and its length, in bytes.
  */
-int ThreadSetName(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadSetName(const Syscall::Args *args, const uintptr_t number);
 /// Resumes the given paused thread.
-int ThreadResume(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadResume(const Syscall::Args *args, const uintptr_t number);
 /// Waits for the specified thread to terminate.
-int ThreadJoin(const Syscall::Args *args, const uintptr_t number);
+intptr_t ThreadJoin(const Syscall::Args *args, const uintptr_t number);
 
 
 /**
  * Implements the create task syscall.
  */
-int TaskCreate(const Syscall::Args *args, const uintptr_t number);
+intptr_t TaskCreate(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Terminates a task, setting its exit code.
@@ -87,7 +87,7 @@ int TaskCreate(const Syscall::Args *args, const uintptr_t number);
  * The first argument is the task handle; if zero, the current task is terminated. The second
  * argument specifies the return code.
  */
-int TaskTerminate(const Syscall::Args *args, const uintptr_t number);
+intptr_t TaskTerminate(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Implements the "initialize task" syscall.
@@ -96,7 +96,7 @@ int TaskTerminate(const Syscall::Args *args, const uintptr_t number);
  * setting up some kernel structures, then perform a return to userspace to the specified address
  * and stack.
  */
-int TaskInitialize(const Syscall::Args *args, const uintptr_t number);
+intptr_t TaskInitialize(const Syscall::Args *args, const uintptr_t number);
 
 /**
  * Sets the descriptive name of the task.
@@ -104,14 +104,14 @@ int TaskInitialize(const Syscall::Args *args, const uintptr_t number);
  * The first argument is a task handle (or zero for current task,) followed by the userspace
  * address of the name and its length, in bytes.
  */
-int TaskSetName(const Syscall::Args *args, const uintptr_t number);
+intptr_t TaskSetName(const Syscall::Args *args, const uintptr_t number);
 
-int TaskDbgOut(const Syscall::Args *args, const uintptr_t number);
+intptr_t TaskDbgOut(const Syscall::Args *args, const uintptr_t number);
 
 /// Installs an IRQ handler
-int IrqHandlerInstall(const Syscall::Args *args, const uintptr_t number);
+intptr_t IrqHandlerInstall(const Syscall::Args *args, const uintptr_t number);
 /// Removes an IRQ handler
-int IrqHandlerRemove(const Syscall::Args *args, const uintptr_t number);
+intptr_t IrqHandlerRemove(const Syscall::Args *args, const uintptr_t number);
 
 }
